refactor: update mapper imports from types to schema
